+++ src/app/cart-module/cart-module.module.ts
@@ -2,6 +2,7 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { CartListComponent } from './cart/component/cart-list/cart-list.component';
 import { CartService } from './cart/service/cart.service';
+import { SharedModuleModule } from '../shared-module/shared-module.module';
 
 @NgModule({
   declarations: [
@@ -10,7 +11,8 @@ import { CartService } from './cart/service/cart.service';
     CartService
   ],
   imports: [
-    CommonModule
+    CommonModule,
+    SharedModuleModule
   ],
   exports: [CartListComponent]
 })