+++ changelog.md
@@ -27,3 +27,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added logic based on Inputs\Outputs
 - Added card items count and cost calculation to CadService
 
+
+## Task2 [1.0.0] - 2020-02-05
+- Added Highlight directive to shared module
+- added title using ViewChild
+- playing with # binding and key events
+